Slightly faster chaining by avoiding two ifs in the inner loop
Whenever we start working on a new anchor that has a different ref_id than
the previous one or that is more than max_ref_gap away from the previous
one, we can directly update the index of the earliest anchor to look at
(lookup_end) to the current one.

Also, the outer loop can start from 1 instead of 0 because the first anchor
cannot have a predecessor.